working version of modal with updated syntax
