+++ package.json
@@ -59,6 +59,7 @@
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
     "@mui/material": "^5.10.3",
+    "@xstate/react": "^3.0.1",
     "hono": "^2.1.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
